+++ processor/src/test/fixtures/unresolved/com/example/autodetect/AppInjector.java
@@ -1,0 +1,10 @@
+package com.example.autodetect;
+
+import sting.Injector;
+
+@Injector
+interface AppInjector
+{
+  LibModel getLibModel();
+}
+